Consolidate Graylog Collector resources
Fixes #1273
