+++ graylog2-server/src/main/java/org/graylog2/rest/resources/system/collector/CollectorResource.java
@@ -17,6 +17,8 @@
 package org.graylog2.rest.resources.system.collector;
 
 import com.codahale.metrics.annotation.Timed;
+import com.github.joschi.jadconfig.util.Duration;
+import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Function;
 import com.google.common.primitives.Ints;
 import com.wordnik.swagger.annotations.Api;
@@ -26,60 +28,52 @@ import com.wordnik.swagger.annotations.ApiResponse;
 import com.wordnik.swagger.annotations.ApiResponses;
 import org.apache.shiro.authz.annotation.RequiresAuthentication;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
-import org.graylog2.Configuration;
 import org.graylog2.collectors.Collector;
 import org.graylog2.collectors.CollectorService;
 import org.graylog2.collectors.Collectors;
-import org.graylog2.database.NotFoundException;
+import org.graylog2.rest.models.collector.requests.CollectorRegistrationRequest;
 import org.graylog2.rest.models.collector.responses.CollectorList;
 import org.graylog2.rest.models.collector.responses.CollectorSummary;
 import org.graylog2.shared.rest.resources.RestResource;
 import org.graylog2.shared.security.RestPermissions;
+import org.hibernate.validator.constraints.NotEmpty;
 import org.joda.time.DateTime;
 
 import javax.inject.Inject;
+import javax.inject.Named;
+import javax.validation.Valid;
+import javax.validation.constraints.NotNull;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.GET;
+import javax.ws.rs.HeaderParam;
+import javax.ws.rs.NotFoundException;
+import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
+import javax.ws.rs.core.Response;
 import java.util.List;
 
 @Api(value = "System/Collectors", description = "Management of Graylog Collectors.")
 @Path("/system/collectors")
 @Consumes(MediaType.APPLICATION_JSON)
 @Produces(MediaType.APPLICATION_JSON)
-@RequiresAuthentication
-@RequiresPermissions(RestPermissions.COLLECTORS_READ)
 public class CollectorResource extends RestResource {
     private final CollectorService collectorService;
     private final LostCollectorFunction lostCollectorFunction;
 
-    public class LostCollectorFunction implements Function<Collector, Boolean> {
-        private final long timeOutInSeconds;
-
-        @Inject
-        public LostCollectorFunction(long timeOutInSeconds) {
-            this.timeOutInSeconds = timeOutInSeconds;
-        }
-
-        @Override
-        public Boolean apply(Collector collector) {
-            final DateTime threshold = DateTime.now().minusSeconds(Ints.checkedCast(timeOutInSeconds));
-            return collector.getLastSeen().isAfter(threshold);
-        }
-    }
-
     @Inject
-    public CollectorResource(CollectorService collectorService, Configuration config) {
+    public CollectorResource(CollectorService collectorService, @Named("collector_inactive_threshold") Duration inactiveThreshold) {
         this.collectorService = collectorService;
-        this.lostCollectorFunction = new LostCollectorFunction(config.getCollectorInactiveThreshold().toSeconds());
+        this.lostCollectorFunction = new LostCollectorFunction(inactiveThreshold.toSeconds());
     }
 
     @GET
     @Timed
     @ApiOperation(value = "Lists all existing collector registrations")
+    @RequiresAuthentication
+    @RequiresPermissions(RestPermissions.COLLECTORS_READ)
     public CollectorList list() {
         final List<Collector> collectors = collectorService.all();
         final List<CollectorSummary> collectorSummaries = Collectors.toSummaryList(collectors, lostCollectorFunction);
@@ -93,13 +87,51 @@ public class CollectorResource extends RestResource {
     @ApiResponses(value = {
             @ApiResponse(code = 404, message = "No collector with the specified id exists")
     })
+    @RequiresAuthentication
+    @RequiresPermissions(RestPermissions.COLLECTORS_READ)
     public CollectorSummary get(@ApiParam(name = "collectorId", required = true)
-                            @PathParam("collectorId") String collectorId) throws NotFoundException {
+                                @PathParam("collectorId") @NotEmpty String collectorId) {
         final Collector collector = collectorService.findById(collectorId);
-
-        if (collector != null)
+        if (collector != null) {
             return collector.toSummary(lostCollectorFunction);
-        else
+        } else {
             throw new NotFoundException("Collector <" + collectorId + "> not found!");
+        }
+    }
+
+    @PUT
+    @Timed
+    @Path("/{collectorId}")
+    @ApiOperation(value = "Create/update an collector registration",
+            notes = "This is a stateless method which upserts a collector registration")
+    @ApiResponses(value = {
+            @ApiResponse(code = 400, message = "The supplied request is not valid.")
+    })
+    public Response register(@ApiParam(name = "collectorId", value = "The collector id this collector is registering as.", required = true)
+                             @PathParam("collectorId") @NotEmpty String collectorId,
+                             @ApiParam(name = "JSON body", required = true)
+                             @Valid @NotNull CollectorRegistrationRequest request,
+                             @HeaderParam(value = "X-Graylog-Collector-Version") @NotEmpty String collectorVersion) {
+        final Collector collector = collectorService.fromRequest(collectorId, request, collectorVersion);
+
+        collectorService.save(collector);
+
+        return Response.accepted().build();
+    }
+
+    @VisibleForTesting
+    protected static class LostCollectorFunction implements Function<Collector, Boolean> {
+        private final long timeOutInSeconds;
+
+        @Inject
+        public LostCollectorFunction(long timeOutInSeconds) {
+            this.timeOutInSeconds = timeOutInSeconds;
+        }
+
+        @Override
+        public Boolean apply(Collector collector) {
+            final DateTime threshold = DateTime.now().minusSeconds(Ints.saturatedCast(timeOutInSeconds));
+            return collector.getLastSeen().isAfter(threshold);
+        }
     }
 }
